requirements.txt
Django==4.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
redis==4.5.5
django-redis==5.2.0
celery==5.2.7
django-environ==0.10.0
whitenoise==6.4.0
